level in own class
